+++ core/src/com/ubuntuserver/frontend/Logic/Logic_SelectBox.java
@@ -110,7 +110,7 @@ public class Logic_SelectBox {
 		int startIndex = activeGraph.dataModel.dataClipStart;
 		
 		//allocate space for the data that will be displayed
-		int[] tempData = new int[(endIndex - startIndex) + 1];
+		float[] tempData = new float[(endIndex - startIndex) + 1];
